Fix stop clusters script (#1335)

+++ dev/stop-clusters.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 set -xuo pipefail
 
-kind delete cluster --name battery || true
-k3d cluster delete || true
+if command -v kind; then
+  kind delete cluster --name battery || true
+fi
+
+if command -v k3d; then
+  k3d cluster delete || true
+fi
